requirements.txt
pandas
re
nltk
vaderSentiment
matplotlib
fpdf2
pyfiglet
beaupy
